requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
django-filter==25.2
django-restframework==0.0.1
django-unfold==0.74.1
djangorestframework==3.16.1
dnspython==2.8.0
eventlet==0.40.4
greenlet==3.3.0
idna==3.11
psycopg2-binary==2.9.11
redis==7.1.0
requests==2.32.5
sqlparse==0.5.5
tasks-py==1.0.3
urllib3==2.6.2
